Update pybind for HB and THB file, remove gsHBSpline_.cpp
